refactor: 消除 bugprone-narrowing-conversions / bugprone-branch-clone (B7)
narrowing (7 处)：YuxError._line / setLineNumber / getLineNumber 与 Diagnostic.line
改 size_t；ANTLR Token::getLine() 返回的 size_t 直接流入 YuxError ctor，不再窄化为
int。SourceLocation.line 保留 int（避免 Node._line 全量级联改动），仅在
YuxError::location() 出口处显式 static_cast<int>。

branch-clone (7 处)：
- sema/call_resolve.cpp 2 处重载匹配三分支合并为二分支（第一/第三分支等价）
- compiler/compiler_dyn_vtable.cpp sanitizeForSymbol 三分支合并为三元
- compiler/call/call_fn.cpp 两分支同 push 合并条件
- literal_node.cpp / expr_literal.cpp UTF-8 解码器 `c < 0x80` 单字节分支与
  fallback 等价，删除冗余分支，注释说明兜底语义

xmake test: 181/181 passed.

+++ include/types.h
@@ -121,7 +121,7 @@ struct SourceLocation {
 };
 
 class YuxError : public std::runtime_error {
-    int _line = 0;
+    size_t _line = 0;
     int _col = 0; // 0 表示列未知
     const char* _code = "E0000"; // 指向 ErrorCode 表中的静态字面量
     DiagSeverity _sev = DiagSeverity::Error; // 默认严重等级（来源于 ErrorCodeDef.defaultSev）
@@ -129,18 +129,18 @@ class YuxError : public std::runtime_error {
     vector<string> _notes; // 附加说明（"= note: ..."），可链式 withNote 追加
 
 public:
-    explicit YuxError(const string& msg, int line) : runtime_error(msg), _line(line) {
+    explicit YuxError(const string& msg, size_t line) : runtime_error(msg), _line(line) {
         assert(line > 0 && "YuxError line must be > 0");
     }
 
     template <class... _Types>
-    explicit YuxError(int line, const format_string<_Types...> format, _Types&&... args) : runtime_error(
+    explicit YuxError(size_t line, const format_string<_Types...> format, _Types&&... args) : runtime_error(
         std::vformat(format.get(), std::make_format_args(args...))), _line(line) {
         assert(line > 0 && "YuxError line must be > 0");
     }
 
     template <class... _Types>
-    explicit YuxError(int line, int col, const format_string<_Types...> format, _Types&&... args) : runtime_error(
+    explicit YuxError(size_t line, int col, const format_string<_Types...> format, _Types&&... args) : runtime_error(
         std::vformat(format.get(), std::make_format_args(args...))), _line(line), _col(col) {
         assert(line > 0 && "YuxError line must be > 0");
     }
@@ -153,7 +153,7 @@ public:
 
     // ErrorCode 路径：模板取自 ec.message，code 取自 ec.code
     template <class... _Types>
-    explicit YuxError(int line, int col, const ErrorCodeDef& ec, _Types&&... args) : runtime_error(
+    explicit YuxError(size_t line, int col, const ErrorCodeDef& ec, _Types&&... args) : runtime_error(
         std::vformat(std::string_view(ec.message), std::make_format_args(args...))),
         _line(line), _col(col), _code(ec.code), _sev(ec.defaultSev) {
         assert(line > 0 && "YuxError line must be > 0");
@@ -161,7 +161,7 @@ public:
 
     // 列未知场景的便利重载（驱动层 / 模块层 errorLine）
     template <class... _Types>
-    explicit YuxError(int line, const ErrorCodeDef& ec, _Types&&... args) : runtime_error(
+    explicit YuxError(size_t line, const ErrorCodeDef& ec, _Types&&... args) : runtime_error(
         std::vformat(std::string_view(ec.message), std::make_format_args(args...))),
         _line(line), _col(0), _code(ec.code), _sev(ec.defaultSev) {
         assert(line > 0 && "YuxError line must be > 0");
@@ -174,20 +174,20 @@ public:
         assert(loc.line > 0 && "YuxError line must be > 0");
     }
 
-    void setLineNumber(int line) {
+    void setLineNumber(size_t line) {
         assert(line > 0 && "YuxError line must be > 0");
         _line = line;
     }
 
     void setColumn(int col) { _col = col; }
 
-    [[nodiscard]] int getLineNumber() const {
+    [[nodiscard]] size_t getLineNumber() const {
         return _line;
     }
 
     [[nodiscard]] int getColumn() const { return _col; }
 
-    [[nodiscard]] SourceLocation location() const { return {_line, _col}; }
+    [[nodiscard]] SourceLocation location() const { return {static_cast<int>(_line), _col}; }
 
     [[nodiscard]] const char* getCode() const { return _code; }
